// newcode/libs/ui/desktop/src/lib/models/interfaces/common-utilities.interface.ts
/*
 * $Id$
 * Copyright(c) RIB Software GmbH
 */

/**
 * An interface for Mark Options for checking the Longitude and Latitude.
 */
export interface IMarkOptions {
	/**
	 * The latitude of the Mark options.
	 */
	latitude: number;

	/**
	 * The longitude of the Mark options.
	 */
	longitude: number;
}